Add autoware vision detector example
This example shows ArmNN executing yolov2-tiny running as a ROS node as
part of the Autoware autonomous driving stack. A tutorial is provided
for an object detection demo using this node running on the KITTI
dataset. The tutorial targets HiKey-960. The whole pipeline including
image pre-processing and  decoding is written in c++. A number of
convenience scripts are provided for running the demo written in python
and bash.