COMPMID-7483: Fixed an access after free bug in llama.cpp
This commit fixes an access after free bug affecting the weight
cache feature implemented for llama.cpp. The root cause of the bug
is that the weight cache is freed when the llama context is
destroyed. However, a loaded model can be shared between multiple llama
contexts. This might cause an access after free if any of the remaining
llama contexts attempt to use the shared model.