Add support for Windows on ARM and MacOS to ggml-kleidiai.cpp
This commit adds cpu feature detection for Windows On ARM and MacOS.
For Windows, it uses IsProcessorFeaturePresent API from windows.h
to detect cpu features. However, there is no specific flag for I8MM,
so it will run a small SMMLA program. If illegal instruction error
is catched, it sets isa.i8mm as false. For MacOS, it uses sysctlbyname.

Change-Id: Ifabf3a3d517edc9693cc815d0aba525a8aa4e91d